+++ polkadot-parachain/Cargo.toml
@@ -25,9 +25,9 @@ rococo-parachain-runtime = { path = "../parachains/runtimes/testing/rococo-parac
 shell-runtime = { path = "../parachains/runtimes/starters/shell", optional = true }
 glutton-runtime = { path = "../parachains/runtimes/glutton/glutton-kusama", optional = true }
 seedling-runtime = { path = "../parachains/runtimes/starters/seedling", optional = true }
-asset-hub-polkadot-runtime = { path = "../parachains/runtimes/assets/asset-hub-polkadot" }
-asset-hub-kusama-runtime = { path = "../parachains/runtimes/assets/asset-hub-kusama" }
-asset-hub-westend-runtime = { path = "../parachains/runtimes/assets/asset-hub-westend" }
+asset-hub-polkadot-runtime = { path = "../parachains/runtimes/assets/asset-hub-polkadot", optional = true }
+asset-hub-kusama-runtime = { path = "../parachains/runtimes/assets/asset-hub-kusama", optional = true }
+asset-hub-westend-runtime = { path = "../parachains/runtimes/assets/asset-hub-westend", optional = true }
 collectives-polkadot-runtime = { path = "../parachains/runtimes/collectives/collectives-polkadot" }
 contracts-rococo-runtime = { path = "../parachains/runtimes/contracts/contracts-rococo" }
 bridge-hub-rococo-runtime = { path = "../parachains/runtimes/bridge-hubs/bridge-hub-rococo" }
@@ -103,7 +103,18 @@ wait-timeout = "0.2"
 polkadot-cli = { git = "https://github.com/paritytech/polkadot", branch = "master", features = ["rococo-native"] }
 
 [features]
-default = [ "glutton-runtime", "penpal-runtime", "seedling-runtime", "shell-runtime" ]
+default = [
+	"asset-hub-kusama-runtime",
+	"asset-hub-polkadot-runtime",
+	"asset-hub-westend-runtime",
+	"glutton-runtime",
+	"penpal-runtime",
+	"seedling-runtime",
+	"shell-runtime"
+]
+asset-hub-kusama-runtime = [ "dep:asset-hub-kusama-runtime"]
+asset-hub-polkadot-runtime = [ "dep:asset-hub-polkadot-runtime"]
+asset-hub-westend-runtime = [ "dep:asset-hub-westend-runtime"]
 glutton-runtime = [ "dep:glutton-runtime", "shell-runtime" ]
 penpal-runtime = [ "dep:penpal-runtime" ]
 seedling-runtime = [ "dep:seedling-runtime" ]
