+++ polkadot-parachain/src/command.rs
@@ -149,46 +149,70 @@ fn load_spec(id: &str) -> std::result::Result<Box<dyn ChainSpec>, String> {
 		),
 
 		// -- Asset Hub Polkadot
-		"asset-hub-polkadot-dev" | "statemint-dev" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_polkadot_development_config()),
-		"asset-hub-polkadot-local" | "statemint-local" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_polkadot_local_config()),
+		"asset-hub-polkadot-dev" | "statemint-dev" => ensure_feature!(
+			"asset-hub-polkadot-runtime",
+			Box::new(chain_spec::asset_hub_polkadot::asset_hub_polkadot_development_config())
+		),
+		"asset-hub-polkadot-local" | "statemint-local" => ensure_feature!(
+			"asset-hub-polkadot-runtime",
+			Box::new(chain_spec::asset_hub_polkadot::asset_hub_polkadot_local_config())
+		),
 		// the chain spec as used for generating the upgrade genesis values
-		"asset-hub-polkadot-genesis" | "statemint-genesis" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_polkadot_config()),
+		"asset-hub-polkadot-genesis" | "statemint-genesis" => ensure_feature!(
+			"asset-hub-polkadot-runtime",
+			Box::new(chain_spec::asset_hub_polkadot::asset_hub_polkadot_config())
+		),
 		// the shell-based chain spec as used for syncing
-		"asset-hub-polkadot" | "statemint" =>
-			Box::new(chain_spec::asset_hubs::AssetHubPolkadotChainSpec::from_json_bytes(
+		"asset-hub-polkadot" | "statemint" => ensure_feature!(
+			"asset-hub-polkadot-runtime",
+			Box::new(chain_spec::asset_hub_polkadot::AssetHubPolkadotChainSpec::from_json_bytes(
 				&include_bytes!("../../parachains/chain-specs/asset-hub-polkadot.json")[..],
-			)?),
+			)?)
+		),
 
 		// -- Asset Hub Kusama
-		"asset-hub-kusama-dev" | "statemine-dev" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_kusama_development_config()),
-		"asset-hub-kusama-local" | "statemine-local" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_kusama_local_config()),
+		"asset-hub-kusama-dev" | "statemine-dev" => ensure_feature!(
+			"asset-hub-kusama-runtime",
+			Box::new(chain_spec::asset_hub_kusama::asset_hub_kusama_development_config())
+		),
+		"asset-hub-kusama-local" | "statemine-local" => ensure_feature!(
+			"asset-hub-kusama-runtime",
+			Box::new(chain_spec::asset_hub_kusama::asset_hub_kusama_local_config())
+		),
 		// the chain spec as used for generating the upgrade genesis values
-		"asset-hub-kusama-genesis" | "statemine-genesis" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_kusama_config()),
+		"asset-hub-kusama-genesis" | "statemine-genesis" => ensure_feature!(
+			"asset-hub-kusama-runtime",
+			Box::new(chain_spec::asset_hub_kusama::asset_hub_kusama_config())
+		),
 		// the shell-based chain spec as used for syncing
-		"asset-hub-kusama" | "statemine" =>
-			Box::new(chain_spec::asset_hubs::AssetHubKusamaChainSpec::from_json_bytes(
+		"asset-hub-kusama" | "statemine" => ensure_feature!(
+			"asset-hub-kusama-runtime",
+			Box::new(chain_spec::asset_hub_kusama::AssetHubKusamaChainSpec::from_json_bytes(
 				&include_bytes!("../../parachains/chain-specs/asset-hub-kusama.json")[..],
-			)?),
+			)?)
+		),
 
 		// -- Asset Hub Westend
-		"asset-hub-westend-dev" | "westmint-dev" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_westend_development_config()),
-		"asset-hub-westend-local" | "westmint-local" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_westend_local_config()),
+		"asset-hub-westend-dev" | "westmint-dev" => ensure_feature!(
+			"asset-hub-westend-runtime",
+			Box::new(chain_spec::asset_hub_westend::asset_hub_westend_development_config())
+		),
+		"asset-hub-westend-local" | "westmint-local" => ensure_feature!(
+			"asset-hub-westend-runtime",
+			Box::new(chain_spec::asset_hub_westend::asset_hub_westend_local_config())
+		),
 		// the chain spec as used for generating the upgrade genesis values
-		"asset-hub-westend-genesis" | "westmint-genesis" =>
-			Box::new(chain_spec::asset_hubs::asset_hub_westend_config()),
+		"asset-hub-westend-genesis" | "westmint-genesis" => ensure_feature!(
+			"asset-hub-westend-runtime",
+			Box::new(chain_spec::asset_hub_westend::asset_hub_westend_config())
+		),
 		// the shell-based chain spec as used for syncing
-		"asset-hub-westend" | "westmint" =>
-			Box::new(chain_spec::asset_hubs::AssetHubWestendChainSpec::from_json_bytes(
+		"asset-hub-westend" | "westmint" => ensure_feature!(
+			"asset-hub-westend-runtime",
+			Box::new(chain_spec::asset_hub_westend::AssetHubWestendChainSpec::from_json_bytes(
 				&include_bytes!("../../parachains/chain-specs/asset-hub-westend.json")[..],
-			)?),
+			)?)
+		),
 
 		// -- Polkadot Collectives
 		"collectives-polkadot-dev" =>
@@ -270,14 +294,22 @@ fn load_spec(id: &str) -> std::result::Result<Box<dyn ChainSpec>, String> {
 		path => {
 			let path: PathBuf = path.into();
 			match path.runtime() {
-				Runtime::AssetHubPolkadot => Box::new(
-					chain_spec::asset_hubs::AssetHubPolkadotChainSpec::from_json_file(path)?,
+				Runtime::AssetHubPolkadot =>
+					ensure_feature!("asset-hub-polkadot-runtime", Box::new(
+					chain_spec::asset_hub_polkadot::AssetHubPolkadotChainSpec::from_json_file(path)?,
+				)),
+				Runtime::AssetHubKusama => ensure_feature!(
+					"asset-hub-kusama-runtime",
+					Box::new(
+						chain_spec::asset_hub_kusama::AssetHubKusamaChainSpec::from_json_file(
+							path
+						)?
+					)
 				),
-				Runtime::AssetHubKusama =>
-					Box::new(chain_spec::asset_hubs::AssetHubKusamaChainSpec::from_json_file(path)?),
-				Runtime::AssetHubWestend => Box::new(
-					chain_spec::asset_hubs::AssetHubWestendChainSpec::from_json_file(path)?,
-				),
+				Runtime::AssetHubWestend =>
+					ensure_feature!("asset-hub-westend-runtime", Box::new(
+					chain_spec::asset_hub_westend::AssetHubWestendChainSpec::from_json_file(path)?,
+				)),
 				Runtime::CollectivesPolkadot | Runtime::CollectivesWestend => Box::new(
 					chain_spec::collectives::CollectivesPolkadotChainSpec::from_json_file(path)?,
 				),
@@ -379,9 +411,12 @@ impl SubstrateCli for Cli {
 
 	fn native_runtime_version(chain_spec: &Box<dyn ChainSpec>) -> &'static RuntimeVersion {
 		match chain_spec.runtime() {
-			Runtime::AssetHubPolkadot => &asset_hub_polkadot_runtime::VERSION,
-			Runtime::AssetHubKusama => &asset_hub_kusama_runtime::VERSION,
-			Runtime::AssetHubWestend => &asset_hub_westend_runtime::VERSION,
+			Runtime::AssetHubPolkadot =>
+				ensure_feature!("asset-hub-polkadot-runtime", &asset_hub_polkadot_runtime::VERSION),
+			Runtime::AssetHubKusama =>
+				ensure_feature!("asset-hub-kusama-runtime", &asset_hub_kusama_runtime::VERSION),
+			Runtime::AssetHubWestend =>
+				ensure_feature!("asset-hub-westend-runtime", &asset_hub_westend_runtime::VERSION),
 			Runtime::CollectivesPolkadot | Runtime::CollectivesWestend =>
 				&collectives_polkadot_runtime::VERSION,
 			Runtime::Shell => ensure_feature!("shell-runtime", &shell_runtime::VERSION),
@@ -441,25 +476,31 @@ macro_rules! construct_benchmark_partials {
 	($config:expr, |$partials:ident| $code:expr) => {
 		match $config.chain_spec.runtime() {
 			Runtime::AssetHubKusama => {
-				let $partials = new_partial::<asset_hub_kusama_runtime::RuntimeApi, _>(
-					&$config,
-					crate::service::aura_build_import_queue::<_, AuraId>,
-				)?;
-				$code
+				ensure_feature!("asset-hub-kusama-runtime", {
+					let $partials = new_partial::<asset_hub_kusama_runtime::RuntimeApi, _>(
+						&$config,
+						crate::service::aura_build_import_queue::<_, AuraId>,
+					)?;
+					$code
+				})
 			},
 			Runtime::AssetHubWestend => {
-				let $partials = new_partial::<asset_hub_westend_runtime::RuntimeApi, _>(
-					&$config,
-					crate::service::aura_build_import_queue::<_, AuraId>,
-				)?;
-				$code
+				ensure_feature!("asset-hub-westend-runtime", {
+					let $partials = new_partial::<asset_hub_westend_runtime::RuntimeApi, _>(
+						&$config,
+						crate::service::aura_build_import_queue::<_, AuraId>,
+					)?;
+					$code
+				})
 			},
 			Runtime::AssetHubPolkadot => {
-				let $partials = new_partial::<asset_hub_polkadot_runtime::RuntimeApi, _>(
-					&$config,
-					crate::service::aura_build_import_queue::<_, AssetHubPolkadotAuraId>,
-				)?;
-				$code
+				ensure_feature!("asset-hub-polkadot-runtime", {
+					let $partials = new_partial::<asset_hub_polkadot_runtime::RuntimeApi, _>(
+						&$config,
+						crate::service::aura_build_import_queue::<_, AssetHubPolkadotAuraId>,
+					)?;
+					$code
+				})
 			},
 			Runtime::CollectivesPolkadot | Runtime::CollectivesWestend => {
 				let $partials = new_partial::<collectives_polkadot_runtime::RuntimeApi, _>(
@@ -478,34 +519,34 @@ macro_rules! construct_async_run {
 		let runner = $cli.create_runner($cmd)?;
 		match runner.config().chain_spec.runtime() {
 			Runtime::AssetHubWestend => {
-				runner.async_run(|$config| {
+				::cumulus_client_cli::ensure_feature!("asset-hub-westend-runtime", runner.async_run(|$config| {
 					let $components = new_partial::<asset_hub_westend_runtime::RuntimeApi, _>(
 						&$config,
 						crate::service::aura_build_import_queue::<_, AuraId>,
 					)?;
 					let task_manager = $components.task_manager;
 					{ $( $code )* }.map(|v| (v, task_manager))
-				})
+				}))
 			},
 			Runtime::AssetHubKusama => {
-				runner.async_run(|$config| {
+				::cumulus_client_cli::ensure_feature!("asset-hub-kusama-runtime",runner.async_run(|$config| {
 					let $components = new_partial::<asset_hub_kusama_runtime::RuntimeApi, _>(
 						&$config,
 						crate::service::aura_build_import_queue::<_, AuraId>,
 					)?;
 					let task_manager = $components.task_manager;
 					{ $( $code )* }.map(|v| (v, task_manager))
-				})
+				}))
 			},
 			Runtime::AssetHubPolkadot => {
-				runner.async_run(|$config| {
+				::cumulus_client_cli::ensure_feature!("asset-hub-polkadot-runtime",runner.async_run(|$config| {
 					let $components = new_partial::<asset_hub_polkadot_runtime::RuntimeApi, _>(
 						&$config,
 						crate::service::aura_build_import_queue::<_, AssetHubPolkadotAuraId>,
 					)?;
 					let task_manager = $components.task_manager;
 					{ $( $code )* }.map(|v| (v, task_manager))
-				})
+				}))
 			},
 			Runtime::CollectivesPolkadot | Runtime::CollectivesWestend => {
 				runner.async_run(|$config| {
@@ -716,10 +757,11 @@ pub fn run() -> Result<()> {
 						runner.sync_run(|config| {
 							match config.chain_spec.runtime() {
 							Runtime::AssetHubKusama =>
-								cmd.run::<Block, AssetHubKusamaExecutor>(config),
-							Runtime::AssetHubWestend => cmd.run::<Block, AssetHubWestendExecutor>(config),
+								ensure_feature!("asset-hub-kusama-runtime", cmd.run::<Block, AssetHubKusamaExecutor>(config)),
+							Runtime::AssetHubWestend =>
+								ensure_feature!("asset-hub-westend-runtime", cmd.run::<Block, AssetHubWestendExecutor>(config)),
 							Runtime::AssetHubPolkadot =>
-								cmd.run::<Block, AssetHubPolkadotRuntimeExecutor>(config),
+								ensure_feature!("asset-hub-polkadot-runtime", cmd.run::<Block, AssetHubPolkadotRuntimeExecutor>(config)),
 							Runtime::CollectivesPolkadot | Runtime::CollectivesWestend =>
 								cmd.run::<Block, CollectivesPolkadotRuntimeExecutor>(config),
 							Runtime::BridgeHub(bridge_hub_runtime_type) => match bridge_hub_runtime_type {
@@ -973,27 +1015,27 @@ pub fn run() -> Result<()> {
 				}
 
 				match config.chain_spec.runtime() {
-					Runtime::AssetHubPolkadot => crate::service::start_generic_aura_node::<
+					Runtime::AssetHubPolkadot => ::cumulus_client_cli::ensure_feature!("asset-hub-polkadot-runtime", crate::service::start_generic_aura_node::<
 						asset_hub_polkadot_runtime::RuntimeApi,
 						AssetHubPolkadotAuraId,
 					>(config, polkadot_config, collator_options, id, hwbench)
 					.await
 					.map(|r| r.0)
-					.map_err(Into::into),
-					Runtime::AssetHubKusama => crate::service::start_generic_aura_node::<
+					.map_err(Into::into)),
+					Runtime::AssetHubKusama => ::cumulus_client_cli::ensure_feature!("asset-hub-kusama-runtime", crate::service::start_generic_aura_node::<
 						asset_hub_kusama_runtime::RuntimeApi,
 						AuraId,
 					>(config, polkadot_config, collator_options, id, hwbench)
 					.await
 					.map(|r| r.0)
-					.map_err(Into::into),
-					Runtime::AssetHubWestend => crate::service::start_generic_aura_node::<
+					.map_err(Into::into)),
+					Runtime::AssetHubWestend => ::cumulus_client_cli::ensure_feature!("asset-hub-westend-runtime", crate::service::start_generic_aura_node::<
 						asset_hub_westend_runtime::RuntimeApi,
 						AuraId,
 					>(config, polkadot_config, collator_options, id, hwbench)
 					.await
 					.map(|r| r.0)
-					.map_err(Into::into),
+					.map_err(Into::into)),
 					Runtime::CollectivesPolkadot | Runtime::CollectivesWestend =>
 						crate::service::start_generic_aura_node::<
 							collectives_polkadot_runtime::RuntimeApi,
